app/requirements.txt
# Pan-Tilt application dependencies
pyserial==3.4
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
numpy==1.18.3

